消费额度统计

+++ vipmis-server/src/main/java/com/waiterlong/vipmis/domain/vo/DepositLogVo.java
@@ -6,6 +6,7 @@ import com.waiterlong.vipmis.domain.DepositLog;
 import com.waiterlong.vipmis.utils.AbstractMyBeanUtils;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.Date;
 import java.util.List;
 
@@ -28,7 +29,7 @@ public class DepositLogVo {
 
     private Date time;
 
-    private Float value;
+    private BigDecimal value;
 
     public static DepositLogVo convertDepositLog(DepositLog depositLog) {
         if (null == depositLog) {
